+++ docs/app/components/content/examples/popover/PopoverDismissibleExample.vue
@@ -1,20 +1,18 @@
 <script setup lang="ts">
 import Cross50Icon from '@bitrix24/b24icons-vue/actions/Cross50Icon'
-
-const open = ref(false)
 </script>
 
 <template>
-  <B24Popover v-model:open="open" :dismissible="false" :b24ui="{ content: 'p-4' }">
+  <B24Popover :dismissible="false" :b24ui="{ content: 'p-4' }">
     <B24Button label="Open" />
 
-    <template #content>
+    <template #content="{ close }">
       <div class="flex items-center gap-4 mb-4">
         <h2 class="text-highlighted font-semibold">
           Popover non-dismissible
         </h2>
 
-        <B24Button color="air-tertiary" :icon="Cross50Icon" @click="open = false" />
+        <B24Button color="air-tertiary" :icon="Cross50Icon" @click="close" />
       </div>
 
       <Placeholder class="size-full min-h-48" />