+++ src/theme/modal.ts
@@ -12,20 +12,23 @@
 
 export default {
   slots: {
-    overlay: 'fixed inset-0 bg-[#003366]/20',
+    overlay: 'fixed inset-0',
     content: [
       'light',
       'bg-(--popup-window-background-color)',
-      'fixed',
       'flex flex-col gap-[20px]',
       'focus:outline-none',
       'p-[24px] pt-[20px]'
     ].join(' '),
-    contentWrapper: [
-      'flex flex-col gap-[15px] pt-[4px]'
-    ].join(' '),
+    contentWrapper: 'flex flex-col gap-[15px] pt-[4px]',
     header: 'flex items-start justify-between gap-[6px]',
     wrapper: '',
+    body: 'flex-1 text-(length:--ui-font-size-md) leading-normal',
+    footer: [
+      'flex items-center justify-between gap-[10px]',
+      'border-t border-t-1 border-t-(--ui-color-divider-default)',
+      'pt-[18px]'
+    ].join(' '),
     title: [
       'font-[family-name:var(--ui-font-family-primary)]',
       'text-(--b24ui-typography-label-color)',
@@ -38,13 +41,7 @@ export default {
       'text-(--b24ui-typography-description-color)',
       'text-(length:--ui-font-size-sm)'
     ].join(' '),
-    close: '-mt-[4px]',
-    body: 'flex-1 overflow-y-auto text-(length:--ui-font-size-md) leading-normal',
-    footer: [
-      'flex items-center justify-between gap-[10px]',
-      'border-t border-t-1 border-t-(--ui-color-divider-default)',
-      'pt-[18px]'
-    ].join(' ')
+    close: '-mt-[4px]'
   },
   variants: {
     overlayBlur: {
@@ -64,13 +61,29 @@ export default {
       },
       false: {
         content: [
-          'top-1/2 left-1/2 -translate-x-1/2 -translate-y-1/2',
-          'w-[calc(100vw-2rem)] max-w-[32rem] max-h-[calc(100dvh-2rem)] sm:max-h-[calc(100dvh-4rem)]',
+          // // 'top-1/2 left-1/2 -translate-x-1/2 -translate-y-1/2',
+          'w-[calc(100vw-2rem)] max-w-[32rem]',
+          // // 'max-h-[calc(100dvh-2rem)] sm:max-h-[calc(100dvh-4rem)]',
           'rounded-[calc(var(--popup-window-border-radius)-2px)] shadow-lg'
           // @memo see components/popup.css
           // 'ring ring-(--popup-window-border)'
         ].join(' '),
-        contentWrapper: 'overflow-hidden'
+        contentWrapper: '' // //  overflow-hidden
+      }
+    },
+    overlay: {
+      true: {
+        overlay: 'bg-[#003366]/20'
+      }
+    },
+    scrollable: {
+      true: {
+        overlay: 'overflow-y-auto',
+        content: 'relative'
+      },
+      false: {
+        content: 'fixed',
+        body: 'overflow-y-auto'
       }
     },
     scrollbarThin: {
@@ -79,6 +92,26 @@ export default {
       }
     }
   },
+  compoundVariants: [
+    {
+      scrollable: true,
+      fullscreen: false,
+      class: {
+        overlay: 'grid place-items-center p-4 sm:py-8'
+      }
+    },
+    {
+      scrollable: false,
+      fullscreen: false,
+      class: {
+        content: [
+          'top-1/2 left-1/2 -translate-x-1/2 -translate-y-1/2',
+          'max-h-[calc(100dvh-2rem)] sm:max-h-[calc(100dvh-4rem)]' // //  overflow-hidden
+        ].join(' '),
+        contentWrapper: 'overflow-hidden'
+      }
+    }
+  ],
   defaultVariants: {
     scrollbarThin: true,
     overlayBlur: 'auto'